+++ app/Http/Controllers/Api/V1/ContactController.php
@@ -6,8 +6,6 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\ContactRequest;
 use App\Http\Resources\ContactResource;
 use App\Models\Contact;
-use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 
 class ContactController extends Controller
 {
@@ -18,13 +16,14 @@ class ContactController extends Controller
         if ($contacts->isEmpty()) {
             return response()->json([
                 'status' => 'failed',
-                'message' => 'No se encontraron contactos.',
-            ], 200);
+                'message' => '¡No se Encontraron Contactos!',
+                'data' => [],
+            ], 404);
         }
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Contactos obtenidos correctamente.',
+            'message' => '¡Contactos Encontrados!',
             'data' => $contacts,
         ], 200);
     }
@@ -35,38 +34,68 @@ class ContactController extends Controller
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Usuario creado exitosamente.',
+            'message' => '¡Contacto Añadido Correctamente!',
             'data' => new ContactResource($contact),
         ], 201);
     }
 
+    public function show(Contact $contact)
+    {
+        if (is_null($contact)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Contacto!',
+            ], 404);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => '!Mostrando Datos del Contacto ' . $contact->name . '!',
+            'data' => new ContactResource($contact),
+        ], 200);
+    }
+
     public function update(ContactRequest $request, Contact $contact)
     {
+        if (is_null($contact)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Contacto para Actualziar!',
+            ], 404);
+        }
+
         $contact->update($request->all());
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Contacto actualizado correctamente.',
-            'data' => $contact,
+            'message' => '¡Contacto Actualizado!',
+            'data' => new ContactResource($contact),
         ], 200);
     }
 
     public function destroy(Contact $contact)
     {
+        if (is_null($contact)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Contacto para Eliminar!',
+            ], 404);
+        }
+
         $contact->delete();
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Contacto eliminado correctamente.',
-            'data' => null,
-        ], 200);
+            'message' => '¡Contacto Eliminado!',
+            'data' => $contact,
+        ], 204);
     }
 
     public function error()
     {
         return response()->json([
             'status' => 'error',
-            'message' => 'Ha ocurrido un error.',
+            'message' => '¡Ha Ocurrido un Error con los Métodos del Controlador para Contactos!',
         ], 400);
     }
 }
