+++ app/Http/Controllers/Api/V1/PhoneBeneficiaryController.php
@@ -17,13 +17,14 @@ class PhoneBeneficiaryController extends Controller
         if ($phoneBeneficiaries->isEmpty()) {
             return response()->json([
                 'status' => 'failed',
-                'message' => 'No se encontraron teléfonos para este beneficiario.',
-            ], 200);
+                'message' => '¡No se Encontraron Teléfonos de Beneficiarios!',
+                'data' => [],
+            ], 404);
         }
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Teléfonos de Beneficiario obtenidos correctamente.',
+            'message' => '¡Estos son los Teléfonos de los Beneficiarios!',
             'data' => $phoneBeneficiaries,
         ], 200);
     }
@@ -34,38 +35,68 @@ class PhoneBeneficiaryController extends Controller
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Usuario creado exitosamente.',
+            'message' => '¡Añadido Nuevo Teléfono para el Beneficiario!',
             'data' => new PhoneBeneficiaryResource($phoneBeneficiary),
         ], 201);
     }
 
+    public function show(PhoneBeneficiaryResource $phoneBeneficiary)
+    {
+        if (is_null($phoneBeneficiary)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Teléfono del Beneficiario!',
+            ], 404);
+        }
+
+        return response()->json([
+            'status' => 'success',
+            'message' => '!Teléfono Encontrado!',
+            'data' => new PhoneBeneficiaryResource($phoneBeneficiary),
+        ], 200);
+    }
+
     public function update(PhoneBeneficiaryRequest $request, PhoneBeneficiary $phoneBeneficiary)
     {
+        if (is_null($phoneBeneficiary)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Teléfono del Beneficiario para Actualizar!',
+            ], 404);
+        }
+
         $phoneBeneficiary->update($request->all());
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Teléfono de Beneficiario actualizado correctamente.',
-            'data' => $phoneBeneficiary,
+            'message' => '¡Teléfono de Beneficiario Actualizado!',
+            'data' => new PhoneBeneficiaryResource($phoneBeneficiary),
         ], 200);
     }
 
     public function destroy(PhoneBeneficiary $phoneBeneficiary)
     {
+        if (is_null($phoneBeneficiary)) {
+            return response()->json([
+                'status' => 'failed',
+                'message' => '¡No se ha encontrado el Teléfono del Beneficiario para Eliminar!',
+            ], 404);
+        }
+
         $phoneBeneficiary->delete();
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Teléfono de Beneficiario eliminado correctamente.',
-            'data' => null,
-        ], 200);
+            'message' => '¡Teléfono de Beneficiario Eliminado!',
+            'data' => $phoneBeneficiary,
+        ], 204);
     }
 
-    public function error(Request $request)
+    public function error()
     {
         return response()->json([
             'status' => 'error',
-            'message' => 'Ha ocurrido un error.',
+            'message' => '¡Ha Ocurrido un Error con los Métodos del Controlador para los Teléfonos de Beneficiarios!',
         ], 400);
     }
 }
